+++ extensions/bmrks/src/constants.ts
@@ -3,7 +3,10 @@
  */
 
 // Name of bmrks service
-export const BMRKS_SERVICE_NAME = "bmrks.com";
+export const BMRKS_SERVICE_NAME = "(Basic) Bookmarks";
 
 // Base URL for bmrks service
 export const HOST_URL = "https://bmrks.com";
+
+// Supported browsers for active tab retrieval
+export const SUPPORTED_BROWSERS = ["Arc", "Brave Browser", "Google Chrome", "Safari"];